chore: commit regenerated batch_attachments entity (size_bytes i64→i32)

+++ src-tauri/src/entities/batch_attachments.rs
@@ -16,7 +16,7 @@ pub struct Model {
     pub original_name: String,
     #[sea_orm(column_type = "Text", nullable)]
     pub mime_type: Option<String>,
-    pub size_bytes: i64,
+    pub size_bytes: i32,
     pub created_at: i32,
 }
